# Speedup etchosts delete